feat: Flattens markdown file structure, updates app config
