+++ config/app.json
@@ -6,7 +6,7 @@
     "entrypoints": [
         {
             "key": "@mint-home-page",
-            "path": "mounts/home/index.tsx",
+            "path": "mounts/Home.tsx",
             "page": {
                 "registryKey": "@mint-home-page",
                 "metadata": {
@@ -23,7 +23,7 @@
         },
         {
             "key": "@mint-browse-page",
-            "path": "mounts/browse-articles/index.tsx",
+            "path": "mounts/Browse.tsx",
             "page": {
                 "registryKey": "@mint-browse-page",
                 "metadata": {
@@ -37,7 +37,7 @@
         },
         {
             "key": "@mint-article-bundle",
-            "path": "mounts/articles/index.tsx",
+            "path": "mounts/Post.tsx",
             "page": {
                 "registryKey": "@mint-main-article-page"
             },
@@ -54,8 +54,8 @@
                 "display": {
                     "image": "/doodles/doodles_space_banana.jpeg",
                     "imageAltText": "A doodles nft character emerging from a banana peel in a light blue themed outer space.",
-                    "title": "Using Webworkers To Prefetch Resources Client Side",
-                    "subtitle": "Prefetching your application's resources enhances your end user's experience with your application because it reduces the idle time your user is spending on your site. Here's how it works.",
+                    "title": "PREFETCH RESOURCES WITH WEB WORKERS AND REACT",
+                    "subtitle": "Do you hate your end user? Us too. Remember 2004? Remember myspace and hypertext, you member.",
                     "tags": [
                         {
                             "visible": "React",
@@ -78,7 +78,7 @@
                 "display": {
                     "image": "/doodles/doodles_pink.png",
                     "imageAltText": "A doodles character in a blue shirt against a pink background.",
-                    "title": "Getting Observability Into Webapp Performance",
+                    "title": "CRYSTAL BALL, WHY DOES MY APP SUCK",
                     "subtitle": "Take longer to load your webpage. I'll wait.",
                     "tags": [
                         {
@@ -106,7 +106,7 @@
                 "display": {
                     "image": "/doodles/doodles_cloud.webp",
                     "imageAltText": "A doodle nft character floating on a cloud.",
-                    "title": "using Native HTML5 to Lazy Load Images",
+                    "title": "BE LAZIER: HTML5 LAZY LOADING MEDIA",
                     "subtitle": "Didnt even really need to be an article this one. Too quick.",
                     "tags": [
                         {
@@ -130,8 +130,8 @@
                 "display": {
                     "image": "/doodles/doodles_pink.png",
                     "imageAltText": "A doodles character in a blue shirt against a pink background.",
-                    "title": "Offering an alternatve method for executing code that could fail. Based on the Rust 'Option' concept.",
-                    "subtitle": "There are a number of code smells with the javascript usage of try-catch. Here's a simple alternative that could help standardize your handling of fallable code.",
+                    "title": "[JS PATTERNS]: CODE THAT MIGHT FAIL",
+                    "subtitle": "Its a suggestion, you know. Like do it.",
                     "tags": [
                         {
                             "visible": "Javascript",
@@ -146,8 +146,8 @@
                 "display": {
                     "image": "/doodles/doodles_space_banana.jpeg",
                     "imageAltText": "A doodles nft character emerging from a banana peel in a light blue themed outer space.",
-                    "title": "ANECDOTE: Amazon Games Gotham",
-                    "subtitle": "Done went and made a design system.",
+                    "title": "ANECDOTE: AMAZON GAMES - GOTHAM DESIGN SYSTEM ONE",
+                    "subtitle": "I hate styled components. I've always hated styled components.",
                     "tags": [
                         {
                             "visible": "React",
@@ -166,7 +166,7 @@
                 "display": {
                     "image": "/doodles/doodles_skeleton.webp",
                     "imageAltText": "The react.js logo in a light blue hue.",
-                    "title": "ANECDOTE: Couch Gag 1",
+                    "title": "ANECDOTE: COUCH GAG ONE",
                     "subtitle": "This is my everest.",
                     "tags": [
                         {
@@ -182,8 +182,8 @@
                 "display": {
                     "image": "/doodles/doodles_cloud.webp",
                     "imageAltText": "A doodle nft character floating on a cloud.",
-                    "title": "Build A Cross Window Lighting Synchronizer with the Broadcast Channel, ESM, and Import Maps",
-                    "subtitle": "Broh went wild in this one.",
+                    "title": "ESM + IMPORT MAPS + BROADCAST CHANNEL + POKEMON",
+                    "subtitle": "Went and did a whole lot trying to do a whole little. Why.",
                     "tags": [
                         {
                             "visible": "Broadcast Channel",
@@ -196,6 +196,26 @@
                         {
                             "visible": "Import Maps",
                             "keys": ["import maps", "html5", "javascript", "frameworkless"]
+                        }
+                    ]
+                }
+            },
+            {
+                "key": "@mint/TC39_SIGNALS_1",
+                "contentPath": "data/TC39_SIGNALS_1.md",
+                "display": {
+                    "image": "/doodles/doodles_space_banana.jpeg",
+                    "imageAltText": "A doodle nft character floating on a cloud.",
+                    "title": "TC39 ECMASCRIPT SIGNALS",
+                    "subtitle": "Actually sort of cool.",
+                    "tags": [
+                        {
+                            "visible": "Ecmascript Proposals",
+                            "keys": ["ecmascript"]
+                        },
+                        {
+                            "visible": "Signals",
+                            "keys": ["reactive signals"]
                         }
                     ]
                 }
