+++ package.json
@@ -5,16 +5,13 @@
   "author": "Nick J Galante <nicholasgalante1997@gmail.com>",
   "type": "module",
   "engines": {
-    "node": "18.x"
+    "node": "20.x"
   },
   "packageManager": "pnpm@8.15.0+sha256.fd1eab68a6d403f35cf3259c53780d70b0f14bd74e39da2f917d201f554d8665",
   "scripts": {
     "format": "prettier --write server src webpack build.tsx",
-    
     "summarize-articles": "node scripts/summarize.mjs",
-    
     "test": "echo \"Error: no test specified\" && exit 1",
-    
     "build": "npm-run-all package:clean summarize-articles package:bundleJs package:compileBuildScript package:buildMarkup package:cleanTmp package:copyAssets",
     "package:clean": "rm -rf build",
     "package:compileBuildScript": "webpack --config webpack/webpack.build-script.cjs",
@@ -22,13 +19,11 @@
     "package:cleanTmp": "rm -rf .build-process/",
     "package:bundleJs": "webpack --config webpack/webpack.prod.cjs",
     "package:copyAssets": "cp -R node_modules/heller-2-lite/build/css/* src/styles/* assets/* build/web",
-    
     "dev:HOME": "REQUEST_PAGE=\"@mint-home-page\" webpack serve --config webpack/webpack.dev.cjs",
     "dev:BROWSE": "REQUEST_PAGE=\"@mint-browse-page\" webpack serve --config webpack/webpack.dev.cjs",
     "dev:ARTICLE": "REQUEST_PAGE=\"@mint-main-article-page\" webpack serve --config webpack/webpack.dev.cjs",
     "start": "npx serve build/web",
     "start:server": "node server/pm2.mjs",
-    
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
