+++ src/mounts/Browse.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Browse } from '@/pages/article/browse/Browse';
+import { Browse } from '@/pages/Browse';
 import { render, setupAnalytics } from '@/lib';
 
 render({ Component: Browse, mountingElement: 'app' });
