+++ src/mounts/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, setupAnalytics } from '@/lib';
-import { Home } from '@/pages/home';
+import { Home } from '@/pages/Home';
 
 render({ Component: Home, mountingElement: 'app' });
 setupAnalytics();
