+++ src/mounts/Post.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { getProps, render, setupAnalytics } from '@/lib';
-import { MainArticle, MainArticleProps } from '@/pages/article/main';
+import { MainArticle, MainArticleProps } from '@/pages/Post';
 
 const props = getProps<MainArticleProps>();
 render({ Component: MainArticle, mountingElement: 'app', props });
