--- src/pages/article/browse/BrowseMeta.ts
@@ -1,11 +1,0 @@
-export const BrowsePageClassNames = {
-  Container: 'couch-mint__browse-page-container',
-  CardGridContainer: 'couch-mint__card-container',
-  ActionContainer: 'couch-mint__card-action-container',
-  SearchContainer: 'couch-mint__card-search-container',
-  ButtonChipRow: 'couch-mint__card-button-chip-row',
-  CardWrappingGrid: 'couch-mint__card-wrapping-container',
-  CardAsideColumn: 'couch-mint__card-aside-container',
-  PaginationContainer: 'couch-mint__card-pagination-container',
-  BigCardContainer: 'couch-mint__big-card-container'
-} as const;
