--- src/pages/article/browse/index.ts
@@ -1,1 +1,0 @@
-export * from './Browse';
