--- src/pages/article/main/MainArticleMeta.ts
@@ -1,13 +1,0 @@
-const MainArticleClassNames = {
-  ImageContainer: 'couch-mint__main-article-image-container',
-  MetadataRow: 'couch-mint__main-article-metadata-row',
-  Author: 'couch-mint__main-article-metadata-author couch-gag__highrise-bold',
-  Collection: 'couch-mint__main-article-metadata-collection',
-  Date: 'couch-mint__main-article-metadata-date',
-  ShareButton: 'couch-mint__main-article-share-action',
-  MarkdownContainer: 'couch-mint__main-article-markdown-container',
-  Title: 'couch-mint__main-article-title couch-mint__ls',
-  Subtitle: 'couch-mint__main-article-subtitle'
-} as const;
-
-export { MainArticleClassNames };
