--- src/pages/article/main/index.ts
@@ -1,2 +1,0 @@
-export * from './MainArticle';
-export * from './types';
