--- src/pages/article/main/types.ts
@@ -1,17 +1,0 @@
-type MainArticleProps = {
-  public: {
-    title: string;
-    subtitle: string;
-    image: {
-      src: string;
-      alt: string;
-    };
-    publishing: {
-      date: string;
-      collection: string;
-    };
-    markdown: string;
-  };
-};
-
-export { type MainArticleProps };
