+++ webpack/webpack.common.cjs
@@ -52,6 +52,7 @@ module.exports = {
     new webpack.ProvidePlugin({
       process: 'process/browser'
     }),
-    new webpack.EnvironmentPlugin({ ...process.env })
+    new webpack.EnvironmentPlugin({ ...process.env }),
+    new webpack.ProgressPlugin()
   ]
 };
